vehicle-signal-server-spec/0180-subscribe-success01.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Vehicle Signal Server Spec Test: get method tests</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-helper.js"></script>

<p> Subscribe Test: success case 01</p>
<p> Subscribe Signal.Drivetrain.Transmission.Speed</p>
<p> (* Run VSSS server before starting test) </p>

<h3>
<div id="log"></div>
</h3>

<script>
var t = async_test("Subscribe test: success 01");

var vehicle  = new WebSocket(VISS_URL);

var PH_START            = 0;
var PH_SENT_SUBSCRIBE   = 1;
var PH_SUBSCRIBED       = 2;
var PH_SENT_UNSUBSCRIBE = 3;
var PH_UNSUBSCRIBED     = 4;
var PH_END              = 5;

vehicle.onopen = function() {
  // request a signal
  var reqId = "reqid-12345";
  var path = "Signal.Drivetrain.Transmission.Speed";
  var notifCount = 0;
  var phase = PH_START;
  var subId = null;

  // process messages from the server
  vehicle.onmessage = t.step_func(
    function(event) {
      var msg = JSON.parse(event.data);

      // subscirbe送信後
      if (phase == PH_START) {
        // nothing to do
      } else if (phase == PH_SENT_SUBSCRIBE) {
        // wait for subscribe response
        if (isSubscribeSuccessResponse(reqId, msg)) {
          addLogMessage("subscribeSuccessResponse received. subscribeId = " + msg.subscriptionId);
          subId = msg.subscriptionId
          phase = PH_SUBSCRIBED;
          //assert_true(true, "OK: "+path + ": value = " + msg.value);
        } else if (isSubscribeErrorResponse(reqId, msg)) {
          addLogFailure("Subscribe method failed");
          phase = PH_END;
        } else {
          // subscribe response 以外を受信した場合
          // 基本、このケースはないはず
          // 無視するか、エラーとしてここで終わるか？
          // とりあえずは無視しておく
        }

      } else if (phase == PH_SUBSCRIBED) {
        // wait subscriptionNotification

        if (isSubscriptionNotification(subId, msg)){
          // after received 3 notif, send unsubscribe
          if (notifCount > 4) {
            // do unsubscribe
            addLogMessage("Send unsubscribe request");
            vehicle.send('{"action":"unsubscribe"'
                          +' ,"path":"'+ path +'"'
                          +' ,"requestId":"'+reqId+'"'
                          +'}');
            phase = PH_SENT_UNSUBSCRIBE;
          }
          addLogMessage("subscriptionNotification received. value = " + msg.value);
          notifCount++;
        }

      } else if (phase == PH_SENT_UNSUBSCRIBE) {
        // wait for unsubscribe response
        if (isUnsubscribeSuccess(reqId, subId, msg)) {
          // unsubscibe成功
          addLogMessage("unsubscribeSuccessResponse received");
          phase = PH_UNSUBSCRIBED;
          addLogSuccess("subscribe/unsubscribe successed");
          t.step_timeout(function() {
            assert_true(true, "OK: "+path + ": subscriptionId = " + msg.subId);
            vehicle.close();
            t.done();
          }, TIME_FINISH_WAIT);

        } else if (isUnsubscribeErrorResponse(reqId, subId, msg)) {
          // unsubscribe失敗
          addLogMessage("unsubscribeErrorResponse received");
          // 失敗の場合はもう一度unsubscribeする？
          // server側にsubscribe が残っているはず
          // 当面は無視

        } else {
          // その他ケース
          // 当面は無視
        }

      } else if (phase == PH_UNSUBSCRIBED) {
        // nothing to do

      }

    }
  );

  // in case server doesn't return response
  t.step_timeout(function() {
    addLogFailure("Fishish by timeout");
    t.step_timeout(function() {
      assert_true(false);
      vehicle.close();
      t.done();
    }, TIME_FINISH_WAIT);
  }, TIME_OUT_TIME);

  addLogMessage("<br>Send subscribe request: path="+path);
  vehicle.send('{"action":"subscribe","path":"'+ path +'" ,"requestId":"'+reqId+'"}');
  phase = PH_SENT_SUBSCRIBE;
}
</script>
